+++ [Documentos]/Segunda Entrega/aqua_point.sql
@@ -32,8 +32,8 @@ CREATE TABLE `aqua_points` (
   `point_name` text NOT NULL,
   `point_type` int(11) NOT NULL,
   `local_id` int(11) NOT NULL,
-  `latitude` float NOT NULL,
-  `longitude` float NOT NULL
+  `latitude` double NOT NULL,
+  `longitude` double NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
 
 -- --------------------------------------------------------
